Refactored db module to sub-modules
